Timetable now responds to `timetable.next_visits` RPCs. This took a lot longer than expected... At first, I wasn't even able to get a working connection to a Crossbar router (See issue 49 on Spell's repo). Once I did get a working connection, I was very confused about how Spell propagated messages as they came in through WAMP. It turns out that the messages are sent to the process that called `Spell.connect/2`, which I initially had as the Supervisor of the process that would be listening for new calls (expecting the calls to be dealt with in the same process).
I then tried using `Task` as a receive loop to continuously wait for incoming messages. However, Spell's public API doesn't allow for indefinite waiting (since it would block the process from doing anything else - I'm still not a fan of this, though), so that option was discarded.

After analyzing the rather-unwieldy RPC example that Spell gives, I noticed that `Callee` was using GenServer and trapping the events in a `handle_info` callback (another design I don't like. It feels like an abuse of the GenServer paradigm). I essentially copy pasted the code, stripped out the parts I didn't need, and ended up with a working server.

And that's where things are now. The server doesn't know to stay open yet, so keeping it running requires `iex -S mix`. After the server starts, it will respond to `timetable.next_visits` with the response "I got it!".